Rename JMC agent classes and package, remove LocalProbeTemplateService
