+++ src/main/java/io/cryostat/core/jmcagent/CapturedValue.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cryostat.core.agent;
+package io.cryostat.core.jmcagent;
 
 import java.net.URI;
 import java.net.URISyntaxException;
