+++ src/main/java/io/cryostat/core/jmcagent/Event.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.cryostat.core.agent;
+package io.cryostat.core.jmcagent;
 
 import java.util.ArrayList;
 import java.util.List;
