--- src/test/resources/io/cryostat/core/agent/template.xml
@@ -1,30 +1,0 @@
-<jfragent>
-    <!-- Global configuration options -->
-    <config>
-        <!-- This is the prefix to use when generating event class names. -->
-        <classprefix>__JFREvent</classprefix>
-        <!-- Will allow the recording of arrays and object parameters as Strings.
-             This will cause toString to be called for array elements and objects other
-             than strings, which in turn can cause trouble if the toString method is badly
-             implemented. Use with care. -->
-        <allowtostring>true</allowtostring>
-        <!-- Allows converters to be used.
-             See the org.openjdk.jmc.agent.converters package for more information. -->
-        <allowconverter>true</allowconverter>
-    </config>
-    <events>
-        <event id="demo.jfr.test1">
-            <label>DemoEvent</label>
-            <description>Event for the agent plugin demo</description>
-            <path>demo</path>
-            <stacktrace>true</stacktrace>
-            <class>io.cryostat.net.HttpServer</class>
-            <method>
-                <name>isAlive</name>
-                <descriptor>()Z</descriptor>
-            </method>
-            <!-- location {ENTRY, EXIT, WRAP}-->
-            <location>ENTRY</location>
-        </event>
-    </events>
-</jfragent>
